refactor: moved adding events back to event management component
